Sort class members to ensure deterministic builds
In an incremental build, class members are already sorted when they are
read from the sorted serialized data.

However, in a clean build where class members are read directly from the
source files, they are not yet sorted. Therefore, there can be a
difference in the contents of the stub files between a clean build and
the next incremental build, making a subsequent task run unnecessarily.

To fix that, this commit ensures that we always sort the class members.

Bug: KT-40882 (there are actually 2 issues in here; this commit fixes
     the first one)
Test: New DeterministicBuildIT